+++ tests/unit/api_container_test.py
@@ -1211,6 +1211,35 @@ class CreateContainerTest(BaseAPIClientTest):
             args[1]['headers'], {'Content-Type': 'application/json'}
         )
 
+    @requires_api_version('1.25')
+    def test_create_container_with_host_config_cpu_realtime(self):
+        self.client.create_container(
+            'busybox', 'ls', host_config=self.client.create_host_config(
+                cpu_realtime_period=1000000,
+                cpu_realtime_runtime=5000
+            )
+        )
+
+        args = fake_request.call_args
+        self.assertEqual(args[0][1],
+                         url_prefix + 'containers/create')
+
+        self.assertEqual(json.loads(args[1]['data']),
+                         json.loads('''
+                            {"Tty": false, "Image": "busybox",
+                             "Cmd": ["ls"], "AttachStdin": false,
+                             "AttachStderr": true,
+                             "AttachStdout": true, "OpenStdin": false,
+                             "StdinOnce": false,
+                             "NetworkDisabled": false,
+                             "HostConfig": {
+                                "CpuRealtimePeriod": 1000000,
+                                "CpuRealtimeRuntime": 5000,
+                                "NetworkMode": "default"
+                             }}'''))
+        self.assertEqual(args[1]['headers'],
+                         {'Content-Type': 'application/json'})
+
 
 class ContainerTest(BaseAPIClientTest):
     def test_list_containers(self):